Change test specs: Output format is now independent of input format.
While it might be more intuitiv in most use cases, we should
not make too many assumptions. The main goal so far is to
keep it compatible with previous versions as far as possible,
even though we are still pre-1.0. Implementing the old spec
would also be too much work for its gain.