fix: h2 database crash on mid-length records
This issue affected only users of the local PID system, as well
as the test environment (developers).

The H2 database stores the record collection as a single string
(CHARACTER VARYING). This limits the (encoded) record by
default to 255 characters.

The solution was to raise the limit to the maximum, without
enforcing the use of a text based column. Fixes #135