cleanup!: remove Handle REST impl. for now.
BREAKING CHANGE:
The REST implementation for the Handle system was removed.
Reason: We can not test it using usename password and not
reasonably maintain it. It may be added in future with key /
certificate support as it is in the HandleProtocol implementation.
Preferably as soon as there is an official handle client for REST.