Implement meta-resolver.
It will check if we have control over the prefix of the PID we are trying to resolve, and if so, use the configured system as we likely have administrative access and may have a better connection or other advantages.
Otherwise, use a read-only client to the handle system, to make sure we can resolve external PIDs, even if we only create sandboxed ones.